add plugin doc
