+++ .github/workflows/ci.yaml
@@ -28,12 +28,11 @@ jobs:
     name: Build & Test
     runs-on: ubuntu-latest
     steps:
-      - name: Set up 1.12
-        uses: actions/setup-go@v0
+      - name: Set up Go 1.12
+        uses: actions/setup-go@v1
         with:
           go-version: 1.12
         id: go
-
       - name: Check out code into the Go module directory
         uses: actions/checkout@v2
