+++ go_resty/README.md
@@ -1,9 +1,9 @@
-# Go2sky with gear (v1.21.2)
+# Go2sky with go-resty(v2.2.0)
 
 ## Installation
 
 ```bash
-go get -u github.com/SkyAPM/go2sky-plugins/gear
+go get -u github.com/SkyAPM/go2sky-plugins/go_resty
 ```
 
 ## Usage
@@ -14,31 +14,27 @@ import (
 	"log"
 
 	"github.com/SkyAPM/go2sky"
-	gearplugin "github.com/SkyAPM/go2sky-plugins/gear"
+	"github.com/SkyAPM/go2sky-plugins/go_resty"
 	"github.com/SkyAPM/go2sky/reporter"
-	"github.com/teambition/gear"
 )
 
 func main() {
+	// Use gRPC reporter for production
 	re, err := reporter.NewLogReporter()
 	if err != nil {
 		log.Fatalf("new reporter error %v \n", err)
 	}
-
 	defer re.Close()
 
-	tracer, err := go2sky.NewTracer("gear", go2sky.WithReporter(re))
+	tracer, err := go2sky.NewTracer("gin-server", go2sky.WithReporter(re))
 	if err != nil {
 		log.Fatalf("create tracer error %v \n", err)
 	}
 
-	app := gear.New()
-    
-	//Use go2sky middleware with tracing
-	app.Use(gearplugin.Middleware(tracer))
-
+	// create go_resty client
+	client := go_resty.NewGoResty(tracer)
 	// do something
 }
 ```
 
-[See more](example_gear_test.go).
+[See more](example_go_resty_test.go)
