--- resty/README.md
@@ -1,1 +1,0 @@
-# Go2sky with go-resty
